% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{examples}
\alias{examples}
\title{Open pharmaverse examples}
\usage{
examples()
}
\value{
The function is executed for its side-effects
}
\description{
Open pharmaverse examples
}
\examples{
\dontrun{
pharmaverse::examples()
}
}
