% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore}
\alias{explore}
\title{Explore pharmaverse homepage}
\usage{
explore()
}
\value{
The function is executed for its side-effects
}
\description{
Explore pharmaverse homepage
}
\examples{
\dontrun{
pharmaverse::explore()
}
}
