% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{join}
\alias{join}
\title{Join a pharmaverse workspace}
\usage{
join(space)
}
\arguments{
\item{space}{\code{character} Name of the space you would like to join:
\itemize{
\item "github"
\item "slack"
\item "linkedin"
}}
}
\value{
The function is executed for its side-effects
}
\description{
Join a pharmaverse workspace
}
\examples{
\dontrun{
pharmaverse::join("github")
}
}
