% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmaverse-package.R
\docType{package}
\name{pharmaverse-package}
\alias{pharmaverse}
\alias{pharmaverse-package}
\title{pharmaverse: Navigate Pharmaverse}
\description{
The 'pharmaverse' is a set of packages that compose multiple pathways through clinical data generation and reporting in the pharmaceutical industry. Learn more about the 'pharmaverse' at \url{https://www.pharmaverse.org}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pharmaverse/pharmaverse-pkg}
  \item \url{https://pharmaverse.org}
  \item Report bugs at \url{https://github.com/pharmaverse/pharmaverse-pkg/issues}
}

}
\author{
\strong{Maintainer}: Ari Siggaard Knoph \email{aikp@novonordisk.com}

Other contributors:
\itemize{
  \item pharmaverse [copyright holder]
}

}
\keyword{internal}
