% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmaverse-package.R
\docType{package}
\name{pharmaverse-package}
\alias{pharmaverse}
\alias{pharmaverse-package}
\title{pharmaverse: Navigate pharmaverse}
\description{
Easily navigate the pharmaverse packages and communities.
}
\author{
\strong{Maintainer}: Ari Siggaard Knoph \email{aikp@novonordisk.com}

}
\keyword{internal}
