% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmaverse-package.R
\docType{package}
\name{pharmaverse-package}
\alias{pharmaverse}
\alias{pharmaverse-package}
\title{pharmaverse: Navigate 'Pharmaverse'}
\description{
The 'pharmaverse' is a set of packages that compose multiple pathways through clinical data generation and reporting in the pharmaceutical industry. This package is designed to guide users to our work-spaces on 'GitHub', 'Slack' and 'LinkedIn' as well as our website and examples. Learn more about the 'pharmaverse' at \url{https://pharmaverse.org}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pharmaverse/pharmaverse-pkg}
  \item \url{https://pharmaverse.org}
  \item Report bugs at \url{https://github.com/pharmaverse/pharmaverse-pkg/issues}
}

}
\author{
\strong{Maintainer}: Ari Siggaard Knoph \email{aikp@novonordisk.com}

Authors:
\itemize{
  \item Ross Farrugia \email{ross.farrugia@roche.com}
  \item Mike Stackhouse \email{mike.stackhouse@atorusresearch.com} (\href{https://orcid.org/0000-0001-6030-723X}{ORCID})
}

Other contributors:
\itemize{
  \item pharmaverse [copyright holder]
}

}
\keyword{internal}
